11_Assignment/pyproject.toml
[tool.poetry]
name = "manualaize"
version = "0.1.0"
description = "An agentic application to interact with user manuals."
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.2.1"
langchain-openai = "^0.1.7"
langchain-community = "^0.2.1"
langchain-cohere = "^0.1.5"
langgraph = "^0.0.64"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = "^2.7.1"
python-dotenv = "^1.0.1"
tavily-python = "^0.3.3"
qdrant-client = "^1.9.0"
cohere = "^5.5.5"
pypdf = "^4.2.0"
nest_asyncio = "^1.6.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
